Unify package-install primitive: engine seam reuses _pv_install_micropip (single micropip code path)
